fix(tests): resolve 21 consistently failing Playwright tests
- Move `iframe_host` fixture from test_iframe_embedding_security.py to
  security/conftest.py so test_innerhtml_guard_iframe.py can discover it
  (fixes 10 ERRORs)
- Replace `networkidle` wait strategy with `domcontentloaded` in
  test_admin_crud_regression.py to prevent SSE connections from blocking
  indefinitely (fixes 10 ERRORs)
- Add HTMX-aware waits to `clear_filters()` in mcp_registry_page.py,
  waiting for #server-grid re-attachment after each filter reset
  (fixes 1 FAIL race condition)
